Code refactoring: split out job and api code.
- refactor pkg/job/abstract.go into different types of jobs
- refactor cloudwatch and tagging apis into separate packages
- move more structs into pkg/model
- move more prometheus-specific code into promutil
